+++ src/lib/eth.ts
@@ -5,8 +5,7 @@ import { ethers } from 'ethers'
 import { ProviderType } from '@dcl/schemas/dist/dapps/provider-type'
 import { getChainConfiguration } from './chainConfiguration'
 import { isMobile } from './utils'
-import { fetchSingleApplicationFeatures } from '../modules/features/utils'
-import { ApplicationName, FeatureName } from '../modules/features/types'
+import { tryPreviousConnectionWrapper } from './TryPreviousConnectionWrapper'
 
 export type EthereumWindow = Window & {
   ethereum?: Provider & {
@@ -71,59 +70,6 @@ export async function getConnectedProvider(): Promise<Provider | null> {
     return null
   }
 }
-
-class TryPreviousConnectionWrapper {
-  private isWalletConnectV2Enabled: boolean | null = null
-  private fetchIsWalletConnectV2EnabledPromise: Promise<boolean> | null = null
-
-  tryPreviousConnection = async (): ReturnType<
-    typeof connection['tryPreviousConnection']
-  > => {
-    const connectionData = connection.getConnectionData()
-    const providerType = connectionData?.providerType
-
-    if (
-      providerType &&
-      (providerType === ProviderType.WALLET_CONNECT ||
-        providerType === ProviderType.WALLET_CONNECT_V2)
-    ) {
-      const isWalletConnectV2Enabled = await this.getIsWalletConnectV2Enabled()
-
-      if (
-        (isWalletConnectV2Enabled &&
-          providerType === ProviderType.WALLET_CONNECT) ||
-        (!isWalletConnectV2Enabled &&
-          providerType === ProviderType.WALLET_CONNECT_V2)
-      ) {
-        await connection.disconnect()
-      }
-    }
-
-    return connection.tryPreviousConnection()
-  }
-
-  private getIsWalletConnectV2Enabled = async (): Promise<boolean> => {
-    if (this.isWalletConnectV2Enabled !== null) {
-      return this.isWalletConnectV2Enabled
-    }
-
-    if (!this.fetchIsWalletConnectV2EnabledPromise) {
-      this.fetchIsWalletConnectV2EnabledPromise = fetchSingleApplicationFeatures(
-        ApplicationName.DAPPS
-      ).then(result => {
-        this.isWalletConnectV2Enabled = !!result.flags[
-          `${ApplicationName.DAPPS}-${FeatureName.WALLET_CONNECT_V2}`
-        ]
-
-        return this.isWalletConnectV2Enabled
-      })
-    }
-
-    return this.fetchIsWalletConnectV2EnabledPromise
-  }
-}
-
-const tryPreviousConnectionWrapper = new TryPreviousConnectionWrapper()
 
 export async function getSigner(): Promise<ethers.Signer> {
   const provider = await getConnectedProvider()
